pip_requirements.txt
aiflows==0.1.5
pytorch-lightning==1.7.7
torchmetrics==0.11.4
wandb==0.13.3
jupyter==1.0.0
jupyterlab==3.5.0
hydra-core==1.2.0
pre-commit==2.20.0
hydra-colorlog==1.1.0
rich==12.6.0
numpy==1.23.0
pandas==1.5.0
matplotlib==3.5.2
jsonlines==3.1.0
selenium==4.8.3
webdriver_manager==3.8.5
wrapt-timeout-decorator==1.3.12.2
undetected_chromedriver==3.4.6
selenium-wire==5.1.0
selenium_recaptcha_solver==1.8.0
bs4==0.0.1
